Improve landing page (#1208)
* feat/nav: add learn more link

* feat/hero: update subtitle

* feat/trending-markets: update title and subtitle

* feat/market: change synths to spot

* feat/leaderboard: update subtitle

* feat/stats: update title and subtitle

* feat/statistics: update 24h values to all-time values (1/2)

* feat/earnings: update titles and subtitles

* feat/cards: reorder cards and update titles and descriptions

* feat/stats: remove community stats

* feat/features: update title and subtitle

* feat/value-prop-cards: update titles and descriptions

* feat/mobile-card: remove soon tag

* feat/partners: update title

* feat/learn: remove the learn more section

* feat/hero: update landing page screenshot

* feat/statistics: update 24h values to all-time values (2/2)

* fix/earnings: recenter cards and rename icons

* remove optimism logo
