+++ components/PoweredBySynthetix/PoweredBySynthetix.tsx
@@ -3,7 +3,6 @@ import { useTranslation } from 'react-i18next';
 import styled from 'styled-components';
 
 import TextLogo from 'assets/svg/brand/text-logo-white.svg';
-import OptimismIcon from 'assets/svg/partners/optimism.svg';
 import { EXTERNAL_LINKS } from 'constants/links';
 import { ExternalLink, FlexDivColCentered, Paragraph } from 'styles/common';
 
@@ -14,9 +13,6 @@ const PoweredBySynthetix = () => {
 			<Text>{t('common.powered-by')}</Text>
 			<ExternalLink href={EXTERNAL_LINKS.Synthetix.Home}>
 				<TextLogo height="10.5px" />
-			</ExternalLink>
-			<ExternalLink href={EXTERNAL_LINKS.Optimism.Home}>
-				<OptimismIcon height="13px" />
 			</ExternalLink>
 		</Container>
 	);
