+++ sections/homepage/ShortList/ShortList.tsx
@@ -4,7 +4,6 @@ import { useRouter } from 'next/router';
 import React, { useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { CellProps } from 'react-table';
-import { useRecoilValue } from 'recoil';
 import styled from 'styled-components';
 
 import GridSvg from 'assets/svg/app/grid.svg';
@@ -15,9 +14,8 @@ import Table from 'components/Table';
 import ROUTES from 'constants/routes';
 import useENS from 'hooks/useENS';
 import { FuturesStat } from 'queries/futures/types';
-import useGetFuturesDailyTradeStats from 'queries/futures/useGetFuturesDailyTradeStats';
+import useGetFuturesCumulativeStats from 'queries/futures/useGetFuturesCumulativeStats';
 import useGetStats from 'queries/futures/useGetStats';
-import { futuresMarketsState } from 'store/futures';
 import { FlexDivColCentered, FlexDivRow, SmallGoldenHeader, WhiteHeader } from 'styles/common';
 import media, { Media } from 'styles/media';
 import { formatCurrency, formatNumber, zeroBN } from 'utils/formatters/number';
@@ -34,8 +32,6 @@ type Stat = {
 
 const ShortList = () => {
 	const { t } = useTranslation();
-
-	const futuresMarkets = useRecoilValue(futuresMarketsState);
 
 	const statsQuery = useGetStats(true);
 	const stats = useMemo(() => statsQuery.data ?? [], [statsQuery]);
@@ -105,13 +101,7 @@ const ShortList = () => {
 		</>
 	);
 
-	const dailyTradeStats = useGetFuturesDailyTradeStats();
-
-	const openInterest = useMemo(() => {
-		return futuresMarkets
-			.map((market) => market.marketSize.mul(market.price).toNumber())
-			.reduce((total, openInterest) => total + openInterest, 0);
-	}, [futuresMarkets]);
+	const totalTradeStats = useGetFuturesCumulativeStats();
 
 	return (
 		<StackSection>
@@ -266,10 +256,10 @@ const ShortList = () => {
 					<StatsCard>
 						<StatsName>{t('homepage.shortlist.stats.volume')}</StatsName>
 						<StatsValue>
-							{dailyTradeStats.isLoading ? (
+							{totalTradeStats.isLoading ? (
 								<Loader />
 							) : (
-								formatCurrency(Synths.sUSD, dailyTradeStats.data?.totalVolume || zeroBN, {
+								formatCurrency(Synths.sUSD, totalTradeStats.data?.totalVolume || zeroBN, {
 									sign: '$',
 									minDecimals: 0,
 								})
@@ -278,26 +268,17 @@ const ShortList = () => {
 						<GridSvg />
 					</StatsCard>
 					<StatsCard>
-						<StatsName>{t('homepage.shortlist.stats.open-interest')}</StatsName>
-						<StatsValue>
-							{!openInterest ? (
-								<Loader />
-							) : (
-								formatCurrency(Synths.sUSD, openInterest ?? 0, {
-									sign: '$',
-									minDecimals: 0,
-								})
-							)}
-						</StatsValue>
+						<StatsName>{t('homepage.shortlist.stats.traders')}</StatsName>
+						<StatsValue>{statsQuery.isLoading ? <Loader /> : stats.length ?? 0}</StatsValue>
 						<GridSvg />
 					</StatsCard>
 					<StatsCard>
 						<StatsName>{t('homepage.shortlist.stats.trades')}</StatsName>
 						<StatsValue>
-							{dailyTradeStats.isLoading ? (
+							{totalTradeStats.isLoading ? (
 								<Loader />
 							) : (
-								formatNumber(dailyTradeStats.data?.totalTrades ?? 0, { minDecimals: 0 })
+								formatNumber(totalTradeStats.data?.totalTrades ?? 0, { minDecimals: 0 })
 							)}
 						</StatsValue>
 						<GridSvg />
